### 2018-10-15 #30 @william-pfalzgraff
Sure, I would be willing to review this.
