### 2018-10-15 #33 @whedon
[REVIEW]: ThermoState: A State manager for Thermodynamics Courses
**Submitting author:** @bryanwweber (<a href="http://orcid.org/0000-0003-0815-9270">Bryan W. Weber</a>)
**Repository:** <a href="https://github.com/bryanwweber/thermostate" target ="_blank">https://github.com/bryanwweber/thermostate</a>
**Version:** v0.4.1
