### 2018-10-15 #33 @kyleniemeyer
👋 @bryanwweber @william-pfalzgraff @scls19fr the review will take place in this issue. Please note the two reviewer checklists above, and you can leave additional review comments here.

You can also leave specific fixes/suggestions/issues as issues in the repo of the submission, just make sure to either mention *this* issue in those or vice versa so we can see the connection.
