### 2018-10-15 bryanwweber/thermostate#11 @bryanwweber
Several examples need NumPy installed to build on Travis
See https://bryanwweber.github.io/thermostate/air-standard-brayton-cycle-example.html
