### 2018-10-15 bryanwweber/thermostate#12 @bryanwweber
Several warning blocks in the tutorial have duplicated colons
See https://bryanwweber.github.io/thermostate/Tutorial.html#ThermoState
