### 2018-10-16 #33 @bryanwweber
Hi @william-pfalzgraff @scls19fr Thank you for agreeing to review this! In looking at the review issue here, I realized there are no installation instructions in the docs... The easiest way is to install with conda or pip, either of which should work. I think you have to use Python 3.5 or 3.6 because CoolProp is only available for those versions; with pip it has to be 3.5, with conda, you can use my CoolProp package for 3.6 with the following command (that should automatically install CoolProp):

```conda install -c bryanwweber thermostate```

You also need a channel that has pint if you're using conda; I'd recommend `conda-forge`. Please let me know if you need help installing so you can check the functionality of the software, and I'll certainly update the docs as part of this review.

Also to be fixed during this review: bryanwweber/thermostate#10 bryanwweber/thermostate#11 bryanwweber/thermostate#12
