### 2018-10-17 #33 @scls19fr
Hi @bryanwweber ,

Thanks for noticing the lack of install instructions.
Fixing doc could also be a nice improvement that you can probably also achieve.

This project relies on previous work about Python for scientists (Python for Scientific Computing, Python for Scientists and Engineers...) . Maybe you could also cite Matplotlib and Numpy at the very least.
Citing Pint may be more complicated for now according https://github.com/hgrecco/pint/issues/717 but citing it like you did is probably not an issue.
Some of your examples (in the `docs` folder are using IPython) maybe you can also cite it.
You can probably fix all that before going beyond.

About "Statement of need", my feeling is quite mixed, I noticed that CoolProp have a "State" API also, so I wonder why such a package is not include directly into CoolProp?
But maybe this kind of question is out of the scope of this review.
Anyway giving an usage example with non SI units with both libraries could give readers an interest in using Thermostate instead of using directly Coolprop.

Best regards
