### 2018-10-17 bryanwweber/thermostate#13 @william-pfalzgraff
Can't install thermostate using pip and python 3.6
I'm not able to install thermostate using pip. In particular, when I try to install it I get this error message:

$ pip install thermostate
Collecting thermostate
  Downloading https://files.pythonhosted.org/packages/b0/2a/f19e8bd5556565ce12360e7bdde1765927e5d8a192137b94be5174daf6db/thermostate-0.4.2-py3-none-any.whl
Collecting coolprop<6.2,>=6.1.0 (from thermostate)
  Could not find a version that satisfies the requirement coolprop<6.2,>=6.1.0 (from thermostate) (from versions: 2.0.1, 2.0.2, 2.0.4, 2.0.5, 2.0.6, 2.1, 2.2.0, 2.2.1, 2.2.2, 2.2.3, 2.2.4, 2.2.5, 2.3, 2.4, 2.5, 3.0.1, 3.1, 3.1.1, 3.1.2, 3.2.0, 3.3.0, 4.0.0b0, 4.0.0, 4.1.0, 4.1.1, 4.1.2, 4.2.0, 4.2.2, 4.2.3, 4.2.4, 4.2.5, 5.0.0, 5.0.1, 5.0.2, 5.0.3, 5.0.4, 5.0.5, 5.0.6, 5.0.7, 5.0.8, 5.1.0, 5.1.1, 5.1.2, 6.0.0)
No matching distribution found for coolprop<6.2,>=6.1.0 (from thermostate)

It looks like there is an issue with the coolprop dependency.  (I am able to install coolprop using pip, but doing so does not resolve the problem).

If helpful, I am running Anaconda python 3.6.
