### 2018-10-17 #33 @william-pfalzgraff
Hi @bryanwweber, thanks for the install instructions.  Having detailed install instructions directly int he readme for thermostate would be very helpful.

Do I understand correctly that pip can _only_ be used if I am using python 3.5, and not python 3.6?  [Link to issue](https://github.com/bryanwweber/thermostate/issues/13).
