### 2018-10-17 #33 @bryanwweber
@william-pfalzgraff I think the way to say it is, if you're using pip to install, you *must* be using Python 3.5 because CoolProp is only available for Python 3.5 from PyPI. If you're using conda, you can use either Python 3.5 or 3.6. Python 3.7 is supported by ThermoState (because ThermoState itself is pure-Python), but binary installers for CoolProp are not yet available for Python 3.7; you can of course build CoolProp yourself and then separately install ThermoState if you wish.

@scls19fr You said
> You can probably fix all that before going beyond.

Do you want me to fix these citations in the paper before continuing with the review? I think your comment about the statement of need is quite relevant, and I will respond a bit later 😄
