### 2018-10-17 bryanwweber/thermostate#13 @bryanwweber
@william-pfalzgraff You say
> I am able to install coolprop using pip, but doing so does not resolve the problem

Are you able to install CoolProp using pip into this Python 3.6 environment?
