### 2018-10-17 #33 @kyleniemeyer
@scls19fr From my perspective, @bryanwweber can work on the issues you have raised while you work through the remaining review items, and then we can return to the paper itself once he revises it. This is meant to be an iterative continuous process.
