### 2018-10-17 bryanwweber/thermostate#13 @william-pfalzgraff
As far as I know I can install coolprop; my pip command and output for doing so is pasted below.

william-pfalzgraff$ pip install CoolProp
Collecting CoolProp
  Using cached https://files.pythonhosted.org/packages/e6/c8/a206a2cc3cb189b2b84c13e6fcea97a9faafacaa9866dd0e49b42f1b5608/CoolProp-6.0.0.tar.gz
Building wheels for collected packages: CoolProp
  Running setup.py bdist_wheel for CoolProp ... done
  Stored in directory: /Users/williampfalzgraff/Library/Caches/pip/wheels/07/ee/01/4566384c81e7a4d8f901550ea21fe711db8c29e9434aad151f
Successfully built CoolProp
Installing collected packages: CoolProp
Successfully installed CoolProp-6.0.0

After which, for example, "import CoolProp.CoolProp as CP" works fine in python 3.6.5

That said, it looks like I have version 6.0.0, not 6.1.0 as required by thermostate. Will 6.0.0 not work for you?  I do see that the wheels available on PyPI only go up to python 3.5 for CoolProp 6.1.0.

Regarding the issues, to my mind there are two threads of discussion that merit two separate issues.  One is the documentation, including adding installation instructions to the Readme.  The other is making sure that it can be installed using common package managers. If it can't be installed with pip then documenting that fact would be good.  That said, if 6.0.0 would work for you, then is it possible to get the pip install working? I do think making the project as available as possible would be very helpful.

My preference would be to open a separate issue for the ReadMe, but I'll defer to you on that.
