### 2018-10-17 bryanwweber/thermostate#13 @bryanwweber
@william-pfalzgraff Sure, a separate issue would probably make sense to update the docs and README.

As for this issue, I don't know if 6.0.0 will work, I haven't tested it. In any case, the only reason it works is because there's a source distribution available, and you happen to be on a platform where `pip install` finds the C++ compiler necessary for the extension. So for folks on platforms where that's not the case (i.e., Windows), the `pip install` won't really work either. So my preference is just to require the most recent version, and deal with the limitations on pip-installing on Python versions. Do you agree with this line of reasoning?
