### 2018-10-18 2efbabc
Install numpy during build of the documentation
Fixes #11. Specifically install NumPy from conda-forge to bring in
openblas which is much smaller than MKL from defaults.
