### 2018-10-18 7a6a3e9
Fix extra colons in blocks in Tutorial
Fixes #12
