### 2018-10-19 d836c97
Update README with installation instructions and documentation links
Work on #14
