### 2018-10-19 8707820
Add installation instructions to docs
Fixes #14
