+++ docs/index.rst
@@ -3,13 +3,20 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to thermostate's documentation!
+Welcome to ThermoState's documentation!
 =======================================
+
+ThermoState is a Python package that provides easy management of thermodynamic states of simple
+compressible systems. ThermoState relies on `CoolProp <http://www.coolprop.org/>`__ and
+`Pint <https://pint.readthedocs.io>`__ to provide the equations of state and units handling,
+respectively. ThermoState replaces tables that are typically used in engineering courses to
+evaluate properties when solving for the behavior of systems.
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
+   installation
    Tutorial
    examples
    thermostate
