### 2018-10-19 bryanwweber/thermostate#13 @bryanwweber
@william-pfalzgraff My preference is to not encourage people to install outdated versions of their software; there is at least one bug in CoolProp 6.1.0 that I know of (that has been fixed on the master branch) that just bit me on an exam problem 😄 So I'm sure there are other bugs that were fixed between 6.0.0 and 6.1.0. Nonetheless, I think its a good idea for my own understanding to see whether or not the test suite at least passes with 6.0.0, so I'll try that and perhaps make a note in the installation instructions.
