### 2018-10-19 ff909c8
Have to include matplotlib to build the examples in the docs
