### 2018-10-19 bbdcec8
Update paper with NumPy, Matplotlib, Notebook citations
