### 2018-10-19 #33 @scls19fr
Thanks @bryanwweber for this very clear explanation!
LGTM when https://github.com/bryanwweber/thermostate/pull/15/ will be merged
