### 2018-10-19 #33 @william-pfalzgraff
One other minor comment/question from me and then likewise LGTM when  bryanwweber/thermostate#15 is merged.

My comment/question is about community guidelines.  Does posting the project to Github constitute "Clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" in the sense that third parties could do all of these things through Github?  Or is more required, e.g. a statement in the README?  @kyleniemeyer, do you have any thoughts?

If there is not already such a statement, it would be nice to add something to the README about how best to contact the author, how to report bugs, that pull requests are welcomed (if they are) etc.

Otherwise, looks good!
