### 2018-10-20 65fe4a7
Add links to COC and Contributing docs to README
