### 2018-10-20 #33 @bryanwweber
@william-pfalzgraff There are 2 files that cover this: the [Contributing instructions](https://github.com/bryanwweber/thermostate/blob/master/CONTRIBUTING.md) and the [Contributor Covenant Code of Conduct](https://github.com/bryanwweber/thermostate/blob/master/CODE_OF_CONDUCT.md). I've added a note linking to these files into the README! Thank you! 😄

@kyleniemeyer What should I do now? Do I make a new release of the software incorporating these suggestions?
