### 2018-10-20 2cb7128
Remove IPython lexer workaround
The issue described in
https://github.com/spatialaudio/nbsphinx/issues/24 seems to be
resolved by https://github.com/ContinuumIO/anaconda-issues/issues/1430
