### 2018-10-20 341c924
Add CHANGELOG, CONTIBUTING and CODE_OF_CONDUCT to docs
