### 2018-10-20 c4116b3
Don't load the sphinx coverage extension, I don't use it
