### 2018-10-20 b92112a
Use the generic Python 3 intersphinx rather than the version specific
