### 2018-10-20 b99a9c9
Install recommonmark and other packages to build the docs on Travis
