### 2018-10-20 86a07e7
Use pip to install the package on Travis during docs build
