### 2018-10-20 e3d6d94
Test Python 3.7 on Travis
Have to install the nightly version of CoolProp using pip and their
index repo
