### 2018-10-20 0722395
Fix or ignore Markdownlint errors in CHANGELOG
