+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+<!-- markdownlint-disable MD022 MD032 MD024 -->
 
 ## [Unreleased]
 ### Added
