### 2018-10-20 3778660
Update CHANGELOG with changes from JOSE paper work
