### 2018-10-20 bryanwweber/thermostate#13 @bryanwweber
@william-pfalzgraff I checked this and the test suite passes with CoolProp 6.0.0. So I'll say it's officially unsupported, but it works if you need it.
