### 2018-10-20 0b1667c
Actually test with Python 3.7
