### 2018-10-21 bryanwweber/thermostate#13 @william-pfalzgraff
Great, then I think it makes sense to consider this issue closed. Thanks!
