### 2018-10-23 #33 @kyleniemeyer
@bryanwweber a few paper comments/edits:

- why is "manager" lowercase in the paper title? Although I personally prefer sentence-case, it looks like you are using title case, so shouldn't that be capitalized?
- Could you add more-complete affiliation info? (e.g., city, state, country—what you'd typically see in a paper)
- First paragraph in Statement of Need: I believe "students" needs an apostrophe
- Second paragraph in Statement of Need: "don't" -> "do not"
- Third paragraph in Statement of Need: is "utilize" really necessary there? I think "use" would work just fine.
- In the first code snippet, I'm wondering if you should include where `Q_` and `State` come from. I realize that the repo itself has more-complete use examples, but I think those in the paper would be more useful if they had those extra imports.
- For the Oliphant reference, if this is a book I believe more details are needed. If this is an electronic-only reference, then we need a DOI or URL.

Finally, this isn't so much an edit but a general question on the package: why did you choose short single-letter abbreviations for properties rather than writing out the names (e.g., `T, p` vs `temperature, pressure`)? I would think the latter would be more clear (and thus more Pythonic)... the main benefit I can see is that the variables match those when setting using attributes (e.g., `state.Tp = ...`).
