### 2018-10-23 b3ed3d9
Updates to paper from editor suggestions
