+++ paper/paper.bib
@@ -50,7 +50,9 @@
     language = {English},
     author = {Oliphant, Travis E},
     year = {2015},
-    note = {OCLC: 1030608394}
+    note = {OCLC: 1030608394},
+    publisher = {CreateSpace Independent Publishing Platform},
+    edition = 2,
 }
 
 @conference{jupyter_notebook,
