### 2018-10-23 76fe8f2
Updates to paper from editor suggestions
