### 2018-10-23 #33 @kyleniemeyer
@bryanwweber LGTM! And got it, that makes sense.

OK, please archive the entire repo and provide the DOI here.
