+++ thermostate/_version.py
@@ -1,4 +1,4 @@
 """The version of thermohw."""
 
-__version_info__ = (0, 4, 3, 'dev0')  # type: Tuple[int, int, int, str]
+__version_info__ = (0, 5, 0, '')  # type: Tuple[int, int, int, str]
 __version__ = '.'.join([str(v) for v in __version_info__ if str(v)])
