+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- markdownlint-disable MD022 MD032 MD024 -->
 
-## [Unreleased]
+## [0.5.0] - 23-OCT-2018
 ### Added
 - Add JOSE paper
 - Add installation, documentation, code of conduct, and contributing links to README
@@ -138,7 +138,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First Release
 
-[Unreleased]: https://github.com/bryanwweber/thermostate/compare/v0.4.2...HEAD
+[0.5.0]: https://github.com/bryanwweber/thermostate/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/bryanwweber/thermostate/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/bryanwweber/thermostate/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/bryanwweber/thermostate/compare/v0.3.0...v0.4.0
