+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- markdownlint-disable MD022 MD032 MD024 -->
 
+## [Unreleased]
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [0.5.0] - 23-OCT-2018
 ### Added
 - Add JOSE paper
@@ -138,6 +147,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First Release
 
+[Unreleased]: https://github.com/bryanwweber/thermostate/compare/v0.5.0...HEAD
 [0.5.0]: https://github.com/bryanwweber/thermostate/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/bryanwweber/thermostate/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/bryanwweber/thermostate/compare/v0.4.0...v0.4.1
