### 2018-10-23 #33 @bryanwweber
@kyleniemeyer The doi is 10.5281/zenodo.1469771 :tada:
