### 2018-10-23 #33 @labarba
@bryanwweber The paper has you as the sole author, but the Zenodo archive does not match that. It could be that Zenodo grabbed authors automatically from GitHub commits, and you need to manually edit the entry.
