### 2018-10-24 #33 @bryanwweber
Thanks @labarba! I changed the Zenodo entry so I'm the sole author there too. Any other places you can see to improve this submission?
