### 2018-10-24 #33 @labarba
It looks like a great piece of software for teaching thermo and I can't wait to share it with my colleagues!
