### 2018-10-24 #33 @labarba
This paper is now published in JOSE !!!

@william-pfalzgraff, @scls19fr — Thank you for your review and being a part in this adventure in new publishing!
